fix: shift modal outside from ClickOutside component
