+++ datahub-web-react/src/app/policy/ClickOutside.tsx
@@ -1,5 +1,4 @@
 import React, { useEffect, useRef } from 'react';
-import { Modal } from 'antd';
 
 interface OutsideAlerterType {
     children: React.ReactNode;
@@ -10,24 +9,10 @@ interface OutsideAlerterType {
 export default function ClickOutside({ children, onClickOutside, wrapperClassName }: OutsideAlerterType) {
     const wrapperRef = useRef<HTMLDivElement>(null);
 
-    const modalClosePopup = () => {
-        Modal.confirm({
-            title: 'Exit Policy Editor',
-            content: `Are you sure you want to exit policy editor? All changes will be lost`,
-            onOk() {
-                onClickOutside();
-            },
-            onCancel() {},
-            okText: 'Yes',
-            maskClosable: true,
-            closable: true,
-        });
-    };
-
     function handleClickOutside(event) {
         if (wrapperClassName) {
             if (event.target && event.target.classList.contains(wrapperClassName)) {
-                modalClosePopup();
+                onClickOutside();
             }
         } else if (!(wrapperRef.current as HTMLSpanElement).contains((event.target as Node) || null)) {
             onClickOutside();
